Adapted the plugin for the new plugin architecture of XBMC Dharma
Thanks to tnyfst@gmail.com for this patch

Changed the name of the png file to icon.png, removed description.xml and added addon.xml in the new dharma format, upped the version to 1.0.1 since there was already a 1.0.0 posted on the forum, and added a bunch of small features to default.py such as browsing by genre or by most recent, etc.  and also  added a search command.